fix casting and unused variable problems on windows (#255)

+++ src/tpm2_wrap.c
@@ -512,7 +512,7 @@ int wolfTPM2_SelfTest(WOLFTPM2_DEV* dev)
     selfTest.fullTest = YES;
     rc = TPM2_SelfTest(&selfTest);
 #ifdef WOLFTPM_WINAPI
-    if (rc == TPM_E_COMMAND_BLOCKED) { /* 0x80280400 */
+    if (rc == (int)TPM_E_COMMAND_BLOCKED) { /* 0x80280400 */
     #ifdef DEBUG_WOLFTPM
         printf("TPM2_SelfTest not allowed on Windows TBS (err 0x%x)\n", rc);
     #endif
@@ -2510,7 +2510,7 @@ int wolfTPM2_NVStoreKey(WOLFTPM2_DEV* dev, TPM_HANDLE primaryHandle,
     rc = TPM2_EvictControl(&in);
     if (rc != TPM_RC_SUCCESS) {
     #ifdef WOLFTPM_WINAPI
-        if (rc == TPM_E_COMMAND_BLOCKED) { /* 0x80280400 */
+        if (rc == (int)TPM_E_COMMAND_BLOCKED) { /* 0x80280400 */
         #ifdef DEBUG_WOLFTPM
             printf("TPM2_EvictControl (storing key to NV) not allowed on "
                    "Windows TBS (err 0x%x)\n", rc);
